feat(org): Add yasnippets for theorem-like environments
This commit introduces a set of yasnippets to streamline the
creation of labeled, theorem-like environments in Org mode, such as
theorems, definitions, proofs, and propositions.

A new helper function, `my/org-snippet-get-unique-label`, is
introduced to ensure that the labels generated by these snippets are
unique within the buffer.

The smart cross-reference map has also been updated to recognize the
new environment types.